pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "orddc2024"
dependencies = [
        'beautifulsoup4==4.12.3',
        'certifi==2024.7.4',
        'charset-normalizer==3.3.2',
        'contourpy==1.2.1',
        'cycler==0.12.1',
        'dill==0.3.8',
        'ensemble-boxes==1.0.9',
        'filelock==3.15.4',
        'fonttools==4.53.1',
        'fsspec==2024.6.1',
        'gdown==5.2.0',
        'idna==3.7',
        'jinja2==3.1.4',
        'kiwisolver==1.4.5',
        'llvmlite==0.43.0',
        'loguru==0.7.2',
        'markupsafe==2.1.5',
        'matplotlib==3.9.2',
        'mpmath==1.3.0',
        'networkx==3.3',
        'numba==0.60.0',
        'numpy==1.26.4',
        'opencv-python==4.10.0.84',
        'packaging==24.1',
        'pandas==2.2.2',
        'pillow==10.4.0',
        'psutil==6.0.0',
        'py-cpuinfo==9.0.0',
        'pycocotools==2.0.8',
        'pyparsing==3.1.2',
        'pysocks==1.7.1',
        'python-dateutil==2.9.0.post0',
        'pytz==2024.1',
        'pyyaml==6.0.2',
        'requests[socks]==2.32.3',
        'scipy==1.14.1',
        'seaborn==0.13.2',
        'six==1.16.0',
        'soupsieve==2.6',
        'sympy==1.13.2',
        'tabulate==0.9.0',
        'tqdm==4.66.5',
        'typing-extensions==4.12.2',
        'tzdata==2024.1',
        'ultralytics==8.2.79',
        'ultralytics-thop==2.0.5',
        'urllib3==2.2.2',
]
